+++ bash2gitlab/__init__.py
@@ -6,8 +6,8 @@ __all__ = [
     "__version__",
     "__about__",
     # decompile
-    "run_shred_gitlab_file",
-    "run_shred_gitlab_tree",
+    "run_decompile_gitlab_file",
+    "run_decompile_gitlab_tree",
     # clean
     "clean_targets",
     # copy2local
@@ -49,6 +49,7 @@ from bash2gitlab.__about__ import __version__
 from bash2gitlab.commands.clean_all import clean_targets
 from bash2gitlab.commands.clone2local import clone_repository_ssh, fetch_repository_archive
 from bash2gitlab.commands.compile_all import run_compile_all
+from bash2gitlab.commands.decompile_all import run_decompile_gitlab_file, run_decompile_gitlab_tree
 from bash2gitlab.commands.detect_drift import run_detect_drift
 from bash2gitlab.commands.doctor import run_doctor
 from bash2gitlab.commands.graph_all import generate_dependency_graph
@@ -58,7 +59,6 @@ from bash2gitlab.commands.map_commit import run_commit_map
 from bash2gitlab.commands.map_deploy import run_map_deploy
 from bash2gitlab.commands.precommit import PrecommitHookError, install, uninstall
 from bash2gitlab.commands.show_config import run_show_config
-from bash2gitlab.commands.shred_all import run_shred_gitlab_file, run_shred_gitlab_tree
 from bash2gitlab.plugins import get_pm
 from bash2gitlab.utils.logging_config import generate_config
 from bash2gitlab.watch_files import start_watch