+++ examples/shred/shred.sh
@@ -1,3 +1,0 @@
-#! /bin/bash
-set -euo pipefail
-bash2gitlab shred --in-file original/.gitlab-ci.yml --out src